refactor: 도메인형 구조로 패키지 리팩토링
