+++ README.md
@@ -1,7 +1,42 @@
 # 1단계 - 카카오 로그인
 
 ## 1. 카카오 개발자 사이트에서 앱 등록 및 REST API 키 설정
-    REST API 키는 카카오에게 "spring-gift 앱 입니다." 자신을 소개하는 역할
     카카오는 REST API 키를 보고 어떤 앱이 요청하는지 인식
 
-##
+## 2. 카카오 로그인 요청 및 인가코드 수신
+    KakaoLoginController
+
+    KakaoLoginController에서 로그인 요청을 받으면
+    카카오 인가 코드를 요청하는 URL을 생성해서 
+    카카오 로그인 페이지로 redirect
+
+    카카오 로그인 및 동의가 완료되면
+    카카오에게 인가 코드를 받는다.
+
+## 3. 인가 코드를 이용한 액세스 토큰 발급
+    KakaoLoginService
+
+    카카오에게 받은 인가 코드를 가지고 토큰을 발급
+    RestTemplate를 이용
+    
+## 4. 액세스 토큰 이용해 카카오 API 호출 및 사용자 정보 조회
+    액세스 토큰을 이용해 카카오 사용자 정보를 조회해서 사용자를 식별할 수 있는 고유 ID 와 닉네임을 얻는다.
+    (카카오 사용자 이메일을 받아서 사용하고 싶었지만 추가 기능 요청이 필요해서 고유 ID 로 대체)
+    (이메일을 이용하지 못해서 카카오 로그인이 아닌 방법으로 이미 회원가입 되어 있는 회원은 계정이 2개 생기는 한계)
+    (즉, 이메일/비밀번호 회원가입과 카카오 로그인이 별개의 경로임)
+
+## 5. 카카오 고유 ID 를 이용해 회원 조회 또는 자동 회원가입
+    조회한 카카오 고유 ID 로 DB의 Member 테이블 조회 (social_id)
+    존재하면 ? 로그인 성공
+    존재하지 않으면 ? 회원가입 (이메일, 비밀번호는 임의 값으로 처리)
+
+## 6. 로그인 or 회원가입 진행 후 JWT 토큰 생성
+
+## 7. 클라이언트에게 JWT 토큰 전달 및 최종 로그인 완료
+
+## 8. 테스트 코드 작성
+    KakaoLoginControllerTest
+    KakaoLoginServiceTest
+
+## 9. 패키지 구조 보기좋게 리팩토링
+    도메인형 구조
