+++ src/main/java/gift/global/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
