+++ src/main/java/gift/member/policy/EmailPolicy.java
@@ -1,4 +1,4 @@
-package gift.policy;
+package gift.member.policy;
 
 public class EmailPolicy {
