+++ src/main/java/gift/member/policy/PasswordPolicy.java
@@ -1,4 +1,4 @@
-package gift.policy;
+package gift.member.policy;
 
 public class PasswordPolicy {
