+++ src/main/java/gift/option/domain/Option.java
@@ -1,5 +1,6 @@
-package gift.domain;
+package gift.option.domain;
 
+import gift.product.domain.Product;
 import jakarta.persistence.*;
 
 @Entity
