+++ src/main/java/gift/option/repository/OptionJpaRepository.java
@@ -1,6 +1,6 @@
-package gift.repository;
+package gift.option.repository;
 
-import gift.domain.Option;
+import gift.option.domain.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
