+++ src/main/java/gift/product/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package gift.global.exception;
+package gift.product.exception;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(Long id) {
