+++ src/main/java/gift/wish/repository/JdbcWishRepositoryImpl.java
@@ -1,6 +1,6 @@
-package gift.repository;
+package gift.wish.repository;
 
-import gift.domain.WishOld;
+import gift.wish.domain.WishOld;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
